+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 #### 2025-03-07
 * POM mode has been added to CV Programming
+* DCC-EX turnout operation extended to support basic accessory decoders.
 
 #### 2025-03-06
 * Y turnout added.